requirement.txt
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9.1
six==1.17.0
urllib3==2.4.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
Deprecated==1.2.18
deprecation==2.1.0
et_xmlfile==2.0.0
fonttools==4.57.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
img2pdf==0.6.0
kiwisolver==1.4.8
lxml==5.4.0
markdown-it-py==3.0.0
matplotlib==3.10.1
mdurl==0.1.2
numpy==2.2.5
ocrmypdf==16.10.1
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pdfminer.six==20250416
pi_heif==0.22.0
pikepdf==9.7.0
pillow==11.2.1
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
rfc3986==1.5.0
rich==14.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
urllib3==2.4.0
wrapt==1.17.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
Deprecated==1.2.18
deprecation==2.1.0
et_xmlfile==2.0.0
fonttools==4.57.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
img2pdf==0.6.0
kiwisolver==1.4.8
lxml==5.4.0
markdown-it-py==3.0.0
matplotlib==3.10.1
mdurl==0.1.2
numpy==2.2.5
ocrmypdf==16.10.1
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pdfminer.six==20250416
pi_heif==0.22.0
pikepdf==9.7.0
pillow==11.2.1
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
rfc3986==1.5.0
rich==14.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
urllib3==2.4.0
wrapt==1.17.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
Deprecated==1.2.18
deprecation==2.1.0
et_xmlfile==2.0.0
fonttools==4.57.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
img2pdf==0.6.0
kiwisolver==1.4.8
lxml==5.4.0
markdown-it-py==3.0.0
matplotlib==3.10.1
mdurl==0.1.2
numpy==2.2.5
ocrmypdf==16.10.1
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pdfminer.six==20250416
pi_heif==0.22.0
pikepdf==9.7.0
pillow==11.2.1
pluggy==1.5.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
rfc3986==1.5.0
rich==14.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
urllib3==2.4.0
wrapt==1.17.2
